set viewer to latched
Hi, I'm using tiny_slam in ROS melodic to create a map from a rosbag file. The problem I have is that everytime I try to save the map, the `map_saver` just keeps on waiting for any incoming messages on the /map-topic but nothing happens.
I found out that if one sets the viewer in tiny_slam.cpp to latched by adding a `true` it works the way it's meant to, i guess.